Все функции по работе с бд перенесенны в репозиторий, однако возникает ошибка запуска, связаная с зависимостями
